update prop keys

+++ src/main/scala/org/grapheco/regionfs/GlobalSetting.scala
@@ -27,7 +27,7 @@ class GlobalSetting(val props: Properties) {
     withOptions(Map("strong" -> Constants.CONSISTENCY_STRATEGY_STRONG,
       "eventual" -> Constants.CONSISTENCY_STRATEGY_EVENTUAL)).asInt
   lazy val maxRegionMemAlive = conf.get(Constants.PARAMETER_KEY_REGION_MEM_MAX_ALIVE).withDefault(Constants.DEFAULT_REGION_MEM_DIRTY_TIMEOUT).asLong
-  lazy val maxRegionMemEntryCount: Long = conf.get(Constants.PARAMETER_KEY_REGION_MEM_MAX_ENTRY_COUNT).withDefault(Constants.DEFAULT_REGION_MEM_ENTRY_COUNT).asLong
+  lazy val maxRegionMemEntryCount: Int = conf.get(Constants.PARAMETER_KEY_REGION_MEM_MAX_ENTRY_COUNT).withDefault(Constants.DEFAULT_REGION_MEM_ENTRY_COUNT).asInt
   lazy val maxRegionMemSize: Long = conf.get(Constants.PARAMETER_KEY_REGION_MEM_MAX_TOTAL_SIZE).withDefault(Constants.DEFAULT_REGION_MEM_SIZE).asLong
   lazy val minWritableRegions: Int = conf.get(Constants.PARAMETER_KEY_MIN_WRITABLE_REGIONS).withDefault(Constants.DEFAULT_MIN_WRITABLE_REGIONS).asInt
   lazy val replicaNum: Int = conf.get(Constants.PARAMETER_KEY_REPLICA_NUM).withDefault(Constants.DEFAULT_REPLICA_NUM).asInt
